applications: machine_learning: remove flpr nodes to fix missing label
Removed DTS nodes associated with the cpuflpr target as they refer to
the cpuflpr_code_partition label from the cpuapp_rx_partitions node. The
label has been removed in this file due to the modification of the
cpuapp_rx_partitions node. The cpuflpr_vpr and cpuapp_cpuflpr_ipc DTS
nodes are removed to prevent build failures.

Ref: NCSDK-29186

Signed-off-by: Kamil Piszczek <Kamil.Piszczek@nordicsemi.no>